requirements.txt
tweepy==4.14.0
python-dotenv==1.0.0
anthropic==0.31.0
openai==1.3.0
requests==2.31.0
python-dateutil==2.8.2
pydantic==2.5.0
